requirements.txt
Django==1.7
argparse==1.2.1
djangorestframework==2.4.2
psycopg2==2.5.4
wsgiref==0.1.2
